Chatbot/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
fastapi==0.115.6
google-ai-generativelanguage==0.6.10
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
limits==5.5.0
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.4
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
redis==5.2.0
requests==2.32.5
rsa==4.9.1
slowapi==0.1.9
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.32.1
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
